# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Auth & Login API",
    "version": "1.0.0",
    "description": "Secure authentication API using Express and Supabase"
  },

  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],

  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },

  "paths": {

    "/auth/signup": {
      "post": {
        "summary": "Create a new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": [
                  "email",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created"
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },

    "/auth/login": {
      "post": {
        "summary": "Log in user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": [
                  "email",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful"
          },
          "400": {
            "description": "Missing input"
          },
          "401": {
            "description": "Invalid login"
          }
        }
      }
    },

    "/auth/logout": {
      "post": {
        "summary": "Log out user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "Logged out"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },

    "/public/info": {
      "get": {
        "summary": "Get public information",
        "responses": {
          "200": {
            "description": "Public information returned"
          }
        }
      }
    },

    "/protected/profile": {
      "get": {
        "summary": "Get protected profile",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Profile returned"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },

    "/protected/dashboard": {
      "get": {
        "summary": "Get protected dashboard",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Dashboard returned"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    }
  }
}